+++ epan/tvbuff.h
@@ -617,7 +617,7 @@ extern gint tvb_skip_wsp_return(tvbuff_t* tvb, const gint offset);
  * it returns 0 (meaning "equal") and -1 otherwise, otherwise return -1.
  */
 extern gint tvb_strneql(tvbuff_t *tvb, const gint offset, const gchar *str,
-    const gint size);
+    const size_t size);
 
 /**
  * Call g_ascii_strncasecmp after checking if enough chars left, returning
